Fixed an interaction problem bewteen the vaading ckeditor addon and the progress indicator used by the wizard.
With a ckeditor plugin displayed on a wizard step the editor's cursor
would periodically reset to 0,0. 

On a slow connection this would happen even more frequently.

It appears that the old progressindicator was updating the dom.
Unmanaged Dom updates cause ckeditor to loose its cursor position.
An upgrade to the ProgressBar eliminated this problem.

I've updated the css to retain the original style used by the wizard.